Add Screen All Client Mode Box
All correct in v. core ... I forgot that I had to add the exclusive turquoise screens also from the Client Mode box ...

the opening of the wizard links to the possibility within the client mode box to choose a language, this combination makes not only the skin embedding that is in the clientmodebox.py code sufficient ..... the crash is type readSkin XML if it enters in the CMbox module .... this would be because it is not using any of our two exclusive skins ... neither turquoise nor OctFHD ... it warns hains and testers that they are used to using PLi skins.